+++ src/server/Handlers/EpicsHandler.java
@@ -1,22 +1,26 @@
-package Server.Handlers;
+package server.Handlers;
 
-import com.google.gson.*;
+import com.google.gson.Gson;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
-import task.Task;
+import task.Epic;
+import task.Subtask;
 import taskmanager.TaskManager;
-
-import java.io.*;
+import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class TasksHandler extends BaseHttpHandler implements HttpHandler {
+public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
 
     TaskManager manager;
 
-    public TasksHandler(TaskManager manager) {
+    public EpicsHandler(TaskManager manager) {
         this.manager = manager;
     }
 
@@ -33,7 +37,7 @@ public class TasksHandler extends BaseHttpHandler implements HttpHandler {
                 case "GET" -> {
                     switch (pathLength) {
                         case 2 -> {
-                            List<Task> tasks = manager.getSimpleTasks();
+                            List<Epic> tasks = manager.getEpicTasks();
                             if (tasks == null) {
                                 sendNotFound(exchange, "Задача не найдена");
                             }
@@ -45,14 +49,26 @@ public class TasksHandler extends BaseHttpHandler implements HttpHandler {
                             if (id.isEmpty()) {
                                 sendNotFound(exchange, "Задача не найдена");
                             } else {
-                                Task newTask = manager.getSimpleTask(id.get());
+                                Epic newTask = manager.getEpicTask(id.get());
                                 if (newTask == null) {
                                     sendNotFound(exchange, "Задача не найдена");
                                 }
                             }
-                            Task task = manager.getSimpleTask(id.get());
+                            Epic task = manager.getEpicTask(id.get());
                             String taskJson = gson.toJson(task);
                             sendText(exchange, taskJson);
+                        }
+                        case 4 -> {
+                            Optional<Integer> epicId = getTaskId(exchange);
+                            if (epicId.isEmpty()) {
+                                sendNotFound(exchange, "Задача не найдена");
+                            } else {
+                                if (manager.getEpicTask(epicId.get()) != null) {
+                                    ArrayList<Subtask> subtasks = manager.getEpicSubtasks(epicId.get());
+                                    String subtasksJson = gson.toJson(subtasks);
+                                    sendText(exchange, subtasksJson);
+                                }
+                            }
                         }
                     }
                 }
@@ -64,21 +80,19 @@ public class TasksHandler extends BaseHttpHandler implements HttpHandler {
                         sendNotAcceptable(exchange, "Это не Json объект");
                     }
                     JsonObject jsonObject = element.getAsJsonObject();
-                    Task taskFromJson = gson.fromJson(jsonObject, Task.class);
+                    Epic taskFromJson = gson.fromJson(jsonObject, Epic.class);
                     switch (pathLength) {
                         case 2 -> {
-                            Task newTask = manager.addSimpleTask(taskFromJson);
-                            if (newTask == null) {
-                                sendNotAcceptable(exchange, "Задача пересекается по времени");
-                            } else {
-                                exchange.sendResponseHeaders(201, 0);
-                                exchange.close();
-                            }
+                            manager.addEpicTask(taskFromJson);
+
+                            exchange.sendResponseHeaders(201, 0);
+                            exchange.close();
+
                         }
                         case 3 -> {
                             Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
-                                Task task2 = manager.updateSimpleTask(taskFromJson);
+                                Epic task2 = manager.updateEpicTask(taskFromJson);
                                 if (task2 != null) {
                                     exchange.sendResponseHeaders(201, 0);
                                     exchange.close();
@@ -94,15 +108,16 @@ public class TasksHandler extends BaseHttpHandler implements HttpHandler {
                 case "DELETE" -> {
                     switch (pathLength) {
                         case 2 -> {
-                            manager.deleteSimpleTasks();
+                            manager.deleteEpicTasks();
                             exchange.sendResponseHeaders(201, 0);
                             exchange.close();
+
                         }
                         case 3 -> {
                             Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
-                                if (manager.getSimpleTask(id.get()) != null) {
-                                    manager.deleteSimpleTask(id.get());
+                                if (manager.getEpicTask(id.get()) != null) {
+                                    manager.deleteEpic(id.get());
                                     exchange.sendResponseHeaders(201, 0);
                                     exchange.close();
                                 } else {
@@ -120,6 +135,7 @@ public class TasksHandler extends BaseHttpHandler implements HttpHandler {
                     }
                     exchange.close();
                 }
+
             }
         } catch (Exception e) {
             System.out.println("Возникло исключение");
