+++ src/server/Handlers/SubtasksHandler.java
@@ -1,4 +1,4 @@
-package Server.Handlers;
+package server.Handlers;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
@@ -6,21 +6,21 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
-import task.Epic;
 import task.Subtask;
 import taskmanager.TaskManager;
+
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
+public class SubtasksHandler extends BaseHttpHandler implements HttpHandler {
+
 
     TaskManager manager;
 
-    public EpicsHandler(TaskManager manager) {
+    public SubtasksHandler(TaskManager manager) {
         this.manager = manager;
     }
 
@@ -31,13 +31,11 @@ public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
             Gson gson = createGson();
             String method = exchange.getRequestMethod();
             int pathLength = getPathLength(exchange);
-
-
             switch (method) {
                 case "GET" -> {
                     switch (pathLength) {
                         case 2 -> {
-                            List<Epic> tasks = manager.getEpicTasks();
+                            List<Subtask> tasks = manager.getSubtasks();
                             if (tasks == null) {
                                 sendNotFound(exchange, "Задача не найдена");
                             }
@@ -49,26 +47,14 @@ public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
                             if (id.isEmpty()) {
                                 sendNotFound(exchange, "Задача не найдена");
                             } else {
-                                Epic newTask = manager.getEpicTask(id.get());
+                                Subtask newTask = manager.getSubtask(id.get());
                                 if (newTask == null) {
                                     sendNotFound(exchange, "Задача не найдена");
                                 }
                             }
-                            Epic task = manager.getEpicTask(id.get());
+                            Subtask task = manager.getSubtask(id.get());
                             String taskJson = gson.toJson(task);
                             sendText(exchange, taskJson);
-                        }
-                        case 4 -> {
-                            Optional<Integer> epicId = getTaskId(exchange);
-                            if (epicId.isEmpty()) {
-                                sendNotFound(exchange, "Задача не найдена");
-                            } else {
-                                if (manager.getEpicTask(epicId.get()) != null) {
-                                    ArrayList<Subtask> subtasks = manager.getEpicSubtasks(epicId.get());
-                                    String subtasksJson = gson.toJson(subtasks);
-                                    sendText(exchange, subtasksJson);
-                                }
-                            }
                         }
                     }
                 }
@@ -80,19 +66,21 @@ public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
                         sendNotAcceptable(exchange, "Это не Json объект");
                     }
                     JsonObject jsonObject = element.getAsJsonObject();
-                    Epic taskFromJson = gson.fromJson(jsonObject, Epic.class);
+                    Subtask taskFromJson = gson.fromJson(jsonObject, Subtask.class);
                     switch (pathLength) {
                         case 2 -> {
-                            manager.addEpicTask(taskFromJson);
-
-                            exchange.sendResponseHeaders(201, 0);
-                            exchange.close();
-
+                            Subtask newTask = manager.addSubtask(taskFromJson);
+                            if (newTask == null) {
+                                sendNotAcceptable(exchange, "Задача пересекается по времени");
+                            } else {
+                                exchange.sendResponseHeaders(201, 0);
+                                exchange.close();
+                            }
                         }
                         case 3 -> {
                             Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
-                                Epic task2 = manager.updateEpicTask(taskFromJson);
+                                Subtask task2 = manager.updateSubtask(taskFromJson);
                                 if (task2 != null) {
                                     exchange.sendResponseHeaders(201, 0);
                                     exchange.close();
@@ -108,16 +96,15 @@ public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
                 case "DELETE" -> {
                     switch (pathLength) {
                         case 2 -> {
-                            manager.deleteEpicTasks();
+                            manager.deleteSubtasks();
                             exchange.sendResponseHeaders(201, 0);
                             exchange.close();
-
                         }
                         case 3 -> {
                             Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
-                                if (manager.getEpicTask(id.get()) != null) {
-                                    manager.deleteEpic(id.get());
+                                if (manager.getSubtask(id.get()) != null) {
+                                    manager.deleteSubtask(id.get());
                                     exchange.sendResponseHeaders(201, 0);
                                     exchange.close();
                                 } else {
@@ -135,11 +122,11 @@ public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
                     }
                     exchange.close();
                 }
-
             }
         } catch (Exception e) {
             System.out.println("Возникло исключение");
         }
 
     }
+
 }
