+++ test/HandlersTests/EpicsHandlerTest.java
@@ -1,10 +1,10 @@
 package HandlersTests;
 
-import Server.Handlers.Adapters.DurationAdapter;
-import Server.Handlers.Adapters.InstantAdapter;
-import Server.Handlers.Adapters.TaskTypeAdapter;
-import Server.Handlers.BaseHttpHandler;
-import Server.HttpTaskServer;
+import server.Handlers.Adapters.DurationAdapter;
+import server.Handlers.Adapters.InstantAdapter;
+import server.Handlers.Adapters.TaskTypeAdapter;
+import server.Handlers.BaseHttpHandler;
+import server.HttpTaskServer;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpServer;
